[NN-0037] Improve GUI , no big bug

+++ Example/GUI/src/Header Files/mainwindow.h
@@ -73,8 +73,8 @@ private slots:
 
 	void on_spinBoxNeurons_valueChanged(int value);
 	void on_spinBoxImageId_valueChanged(int value);
-	void on_spinBoxLearningRate_valueChanged(int value);
-	void on_spinBoxMomentum_valueChanged(int value);
+	void on_spinBoxLearningRate_valueChanged(double value);
+	void on_spinBoxMomentum_valueChanged(double value);
 	void on_spinBoxTrainingRating_valueChanged(int value);
 
 	void on_pushButtonRemoveLayer_clicked();
